# user-service/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080/
    description: Generated server url
...
paths:
...
  /api/v1/account:
    get:
      tags:
        - Account Service
      summary: Получение всех аккаунтов
      operationId: getAllAccounts
      parameters:
        - name: page
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Successfully
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
    post:
      tags:
        - Account Service
      summary: Создание аккаунта
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDto'
        required: true
      responses:
        '200':
          description: Successfully
        '400':
          description: Bad Request
...
  /api/v1/account/{id}:
    get:
      tags:
        - Account Service
      summary: Получение аккаунта по id
      operationId: getAccountById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDto'
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Account Not Found
...
components:
  schemas:
...
    AccountDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        phone:
          type: string
        photo:
          type: string
        about:
          type: string
        city:
          type: string
        country:
          type: string
        token:
          type: string
        statusCode:
          type: string
          enum:
            - FRIEND
            - REQUEST_TO
            - REQUEST_FROM
            - BLOCKED
            - DECLINED
            - SUBSCRIBED
            - NONE
            - WATCHING
            - REJECTING
        firstName:
          type: string
        lastName:
          type: string
        regDate:
          type: string
          format: date-time
        birthDate:
          type: string
          format: date-time
        messagePermission:
          type: string
        lastOnlineTime:
          type: string
          format: date-time
        isOnline:
          type: boolean
        isBlocked:
          type: boolean
        isDeleted:
          type: boolean
        photoId:
          type: string
        photoName:
          type: string
        role:
          type: string
          enum:
            - USER
            - MODERATOR
            - ADMIN
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
        password:
          type: string
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string